pyproject.toml
[tool.poetry]
name = "electric-kiwi-export"
version = "0.1.0"
description = ""
authors = ["Jono Hill <jono@hillnz.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pyaes = "^1.6.1"
requests = "^2.28.0"
arrow = "^1.2.2"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
